refactor: decode text runs in REST adapter, migrate text transformer to snapshot
Slice 5 of the canonicalize carve. The REST wire override tables
(characterStyleOverrides + styleOverrideTable) are now resolved into
plan-neutral per-line runs by the adapter (rest-text.ts); the core text
transformer only classifies + renders those runs, consuming NodeSnapshot
with no @figma/rest-api-spec import.

- rest-text.ts: ports splitLines/computeRunsForLine/computeDelta/merge
  verbatim from the in-core transformer; decodes delta fills via decodePaint
- rest-paint.ts: decodePaint/decodeEffect extracted here so both the node
  and text adapters share them without an import cycle
- snapshot.ts: adds SnapshotText/SnapshotTextRun/SnapshotTextStyle/
  SnapshotHyperlink and NodeSnapshot.text
- text.ts: buildFormattedText/detectBoldWeight/classifyRun consume decoded
  runs; Figma types removed
- built-in.ts: textExtractor passes NodeSnapshot directly (no Figma cast)

Behavior-preserving: goldens byte-identical, 258 tests pass.
